event version cleanup
