+++ src/core/events/abstract-event-type.ts
@@ -10,6 +10,7 @@ export abstract class AbstractEventType {
     const upcastFn = this.supportedTypes[eventTypeName.eventType];
 
     const event = upcastFn ? upcastFn(eventTypeName) : undefined;
+    // Logger.warn(event);
     if (!event) {
       Logger.warn(`Unsupported event received! eventType: ${eventTypeName.eventType}`);
     }
