+++ src/core/events/organization/deleted/1/organization-deleted.event.ts
@@ -1,7 +1,7 @@
 import { OrganizationEvent } from '../../organization.event';
 
 export class OrganizationDeleted extends OrganizationEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(organizationId: string) {
     super(organizationId, OrganizationDeleted.version);
