+++ src/core/events/sensor/activated/1/sensor-activated.event.ts
@@ -1,7 +1,7 @@
 import { SensorEvent } from '../../sensor.event';
 
 export class SensorActivated extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(sensorId: string) {
     super(sensorId, SensorActivated.version);
