+++ src/core/events/sensor/deactivated/1/sensor-deactivated.event.ts
@@ -1,7 +1,7 @@
 import { SensorEvent } from '../../sensor.event';
 
 export class SensorDeactivated extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(sensorId: string) {
     super(sensorId, SensorDeactivated.version);
