+++ src/core/events/sensor/deleted/1/sensor-deleted.event.ts
@@ -1,7 +1,7 @@
 import { SensorEvent } from '../../sensor.event';
 
 export class SensorDeleted extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(sensorId: string) {
     super(sensorId, SensorDeleted.version);
