+++ src/core/events/sensor/ds-deleted/1/sensor-datastreamdeleted.event.ts
@@ -1,7 +1,7 @@
 import {SensorEvent} from '../../sensor.event';
 
 export class DatastreamDeleted extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   public readonly dataStreamId: string;
