+++ src/core/events/sensor/ownership-transferred/1/sensor-ownershiptransferred.event.ts
@@ -1,7 +1,7 @@
 import { SensorEvent } from '../../sensor.event';
 
 export class SensorOwnershipTransferred extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   public readonly oldOrganizationId: string;
   public readonly newOrganizationId: string;
