+++ src/core/events/sensor/relocated/1/sensor-relocated.event.ts
@@ -1,7 +1,7 @@
 import { SensorEvent } from '../../sensor.event';
 
 export class SensorRelocated extends SensorEvent {
-  static version = '1.0.0';
+  static version = '1';
 
   readonly longitude: number;
   readonly latitude: number;
