+++ src/core/events/sensor/updated/1/sensor-updated.event.ts
@@ -1,16 +1,12 @@
 import { SensorEvent } from '../../sensor.event';
 
-export class SensorRegistered extends SensorEvent {
-  static version = '1.0.1';
+export class SensorUpdated extends SensorEvent {
+  static version = '1';
 
-  readonly organizationId: string;
   readonly name: string;
-  readonly location: number[];
-  readonly baseObjectId: string;
   readonly aim: string;
   readonly description: string;
   readonly manufacturer: string;
-  readonly active: boolean;
   readonly observationArea: Record<string, any>;
   readonly documentationUrl: string;
   readonly theme: string[];
@@ -18,19 +14,14 @@ export class SensorRegistered extends SensorEvent {
   readonly typeName: string;
   readonly typeDetails: Record<string, any>;
 
-  constructor(sensorId: string, organizationId: string, name: string, location: number[], baseObjectId: string,
-              aim: string, description: string, manufacturer: string, active: boolean,
+  constructor(sensorId: string, name: string, aim: string, description: string, manufacturer: string,
               observationArea: Record<string, any>, documentationUrl: string, theme: string[], category: string,
               typeName: string, typeDetails: Record<string, any>) {
-    super(sensorId, SensorRegistered.version);
-    this.organizationId = organizationId;
+    super(sensorId, SensorUpdated.version);
     this.name = name;
-    this.location = location;
-    this.baseObjectId = baseObjectId;
     this.aim = aim;
     this.description = description;
     this.manufacturer = manufacturer;
-    this.active = active;
     this.observationArea = observationArea;
     this.documentationUrl = documentationUrl;
     this.theme = theme;
