+++ src/query/processor/organization-es-listener.ts
@@ -1,41 +1,23 @@
 import { Injectable } from '@nestjs/common';
 import { AbstractEsListener } from './abstract.es.listener';
 import { OrganizationProcessor } from './organization.processor';
-import { organizationEventType } from '../../core/events/organization';
 import { CheckpointService } from '../service/checkpoint/checkpoint.service';
 import { EventStorePublisher } from '../../event-store/event-store.publisher';
-import { SubscriptionExistsException } from '../handler/errors/subscription-exists-exception';
+import { Event as ESEvent } from 'geteventstore-promise';
 import { OrganizationEvent } from '../../core/events/organization/organization.event';
+import { organizationEventType } from '../../core/events/organization';
 
 @Injectable()
 export class OrganizationEsListener extends AbstractEsListener {
     constructor(
         eventStore: EventStorePublisher,
         checkpointService: CheckpointService,
-        private readonly organizationProcessor: OrganizationProcessor,
+        organizationProcessor: OrganizationProcessor,
     ) {
-        super('backend-organization-es', eventStore, checkpointService);
+        super('backend-organization-es', '$ce-organization', eventStore, checkpointService, organizationProcessor);
     }
 
-    async openSubscription(): Promise<void> {
-        if (!this.subscriptionExists()) {
-            const onEvent = async (_, eventMessage) => {
-                const offset = eventMessage.positionEventNumber;
-                const callback = () => this.checkpointService.updateOne({_id: this.checkpointId}, {offset});
-
-                const originSync = eventMessage.metadata && eventMessage.metadata.originSync;
-                const event = organizationEventType.getEvent(eventMessage) as OrganizationEvent;
-                try {
-                    await this.organizationProcessor.process(event, originSync);
-                    await callback();
-                } catch {
-                    await callback();
-                }
-            };
-
-            await this.subscribeToStreamFrom('$ce-organization', onEvent);
-        } else {
-            throw new SubscriptionExistsException();
-        }
+    parseEvent(eventMessage: ESEvent): OrganizationEvent {
+        return organizationEventType.getEvent(eventMessage) as OrganizationEvent;
     }
 }
