+++ src/query/processor/sensor.es.listener.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { SensorProcessor } from './sensor.processor';
-import { sensorEventType } from '../../core/events/sensor';
 import { AbstractEsListener } from './abstract.es.listener';
 import { CheckpointService } from '../service/checkpoint/checkpoint.service';
 import { EventStorePublisher } from '../../event-store/event-store.publisher';
-import { SubscriptionExistsException } from '../handler/errors/subscription-exists-exception';
+import { sensorEventType } from '../../core/events/sensor';
 import { SensorEvent } from '../../core/events/sensor/sensor.event';
+import { Event as ESEvent } from 'geteventstore-promise';
 
 @Injectable()
 export class SensorEsListener extends AbstractEsListener {
@@ -13,29 +13,12 @@ export class SensorEsListener extends AbstractEsListener {
     constructor(
         eventStore: EventStorePublisher,
         checkpointService: CheckpointService,
-        private readonly sensorProcessor: SensorProcessor,
+        sensorProcessor: SensorProcessor,
     ) {
-        super('backend-sensor-es', eventStore, checkpointService);
+        super('backend-sensor-es', '$ce-sensor', eventStore, checkpointService, sensorProcessor);
     }
 
-    async openSubscription(): Promise<void> {
-        if (!this.subscriptionExists()) {
-            const onEvent = async (_, eventMessage) => {
-                const offset = eventMessage.positionEventNumber;
-                const callback = () => this.checkpointService.updateOne({_id: this.checkpointId}, {offset});
-
-                const event = sensorEventType.getEvent(eventMessage) as SensorEvent;
-                try {
-                    await this.sensorProcessor.process(event);
-                    await callback();
-                } catch {
-                    await callback();
-                }
-            };
-
-            await this.subscribeToStreamFrom('$ce-sensor', onEvent);
-        } else {
-            throw new SubscriptionExistsException();
-        }
+    parseEvent(eventMessage: ESEvent): SensorEvent {
+        return sensorEventType.getEvent(eventMessage) as SensorEvent;
     }
 }
